chore: :building_construction: Se realia Reestructuración de la arquitectura
