+++ src/main/java/com/rentaherramientas/tolly/application/dto/auth/LoginResponse.java
@@ -1,4 +1,4 @@
-package com.rentaherramientas.tolly.application.dto;
+package com.rentaherramientas.tolly.application.dto.auth;
 
 import io.swagger.v3.oas.annotations.media.Schema;
