+++ src/main/java/com/rentaherramientas/tolly/application/dto/auth/RefreshTokenRequest.java
@@ -1,4 +1,4 @@
-package com.rentaherramientas.tolly.application.dto;
+package com.rentaherramientas.tolly.application.dto.auth;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import jakarta.validation.constraints.NotBlank;
