+++ src/main/java/com/rentaherramientas/tolly/application/dto/tool/CreateToolRequest.java
@@ -1,0 +1,5 @@
+package com.rentaherramientas.tolly.application.dto.tool;
+
+public class CreateToolRequest {
+    
+}
