+++ src/main/java/com/rentaherramientas/tolly/application/mapper/ToolMapper.java
@@ -3,6 +3,7 @@ package com.rentaherramientas.tolly.application.mapper;
 import com.rentaherramientas.tolly.domain.model.enums.ToolStatus;
 import org.mapstruct.Mapper;
 
+
 @Mapper(componentModel = "spring")
 public interface ToolMapper {
