+++ src/main/java/com/rentaherramientas/tolly/application/usecase/auth/LogoutUseCase.java
@@ -1,4 +1,4 @@
-package com.rentaherramientas.tolly.application.usecase;
+package com.rentaherramientas.tolly.application.usecase.auth;
 
 import com.rentaherramientas.tolly.domain.exceptions.DomainException;
 import com.rentaherramientas.tolly.domain.model.RefreshToken;
